+++ crates/sml-core/src/monomorphize.rs
@@ -1,3 +1,4 @@
+#![allow(dead_code)]
 use super::*;
 
 pub struct Cache<'a> {
